+++ examples/vue-apollo/src/components/ErrorSnackBar.vue
@@ -1,39 +1,27 @@
 <template>
-  <v-snackbar
-    v-model="snack"
-    :timeout="2_000"
-  >
+  <v-snackbar v-model="snack" :timeout="2_000" top>
     {{ error?.message }}
     <template #actions>
-      <v-btn
-        color="blue"
-        variant="text"
-        @click="snack = false"
-      >
-        Close
-      </v-btn>
+      <v-btn color="blue" variant="text" @click="snack = false"> Close </v-btn>
     </template>
   </v-snackbar>
 </template>
 
-<script lang="ts">
-import { defineComponent, PropType, ref, watchEffect } from 'vue'
+<script lang="ts" setup>
+import { PropType, ref, watchEffect } from 'vue'
 
 import { ErrorPayload } from '@nhost/core'
-export default defineComponent({
-  props: {
-    error: Object as PropType<ErrorPayload | null>
-  },
-  setup(props) {
-    const snack = ref(false)
-    
-    watchEffect(() => {
-      if (props.error) {
-        snack.value = true
-      }
-    })
 
-    return { snack }
+const props = defineProps({
+  error: Object as PropType<ErrorPayload | null>
+})
+const snack = ref(false)
+
+watchEffect(() => {
+  if (props.error) {
+    snack.value = true
   }
 })
+
+
 </script>