+++ examples/vue-apollo/vite.config.js
@@ -1,5 +1,5 @@
 import { defineConfig } from 'vite'
-
+import path from 'path'
 import vue from '@vitejs/plugin-vue'
 import vuetify from '@vuetify/vite-plugin'