+++ packages/nhost-js/src/clients/graphql.ts
@@ -35,7 +35,7 @@ export class NhostGraphqlClient {
   }
 
   /**
-   * Use `nhost.graphql.request` to send a GraphQL request. For more serious GraphQL usage in your app we recommend using a GraphQL client such as Apollo Client (https://www.apollographql.com/docs/react).
+   * Use `nhost.graphql.request` to send a GraphQL request. For more serious GraphQL usage we recommend using a GraphQL client such as Apollo Client (https://www.apollographql.com/docs/react).
    *
    * @example
    * ```ts