+++ templates/functions/node/javascript.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.status(200).send(`Nhost, from Javascript, pays it's respects to ${req.query.name}!`);
-};